# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Briefed API",
    "description": "Personal AI email agent. Runs a daily pipeline on the user's Gmail inbox, categorizes and prioritizes messages, and summarizes tech news via the Claude API.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/oauth/gmail/start": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Start Google OAuth authorization-code flow",
        "description": "Generate PKCE + state, set the cookie, redirect to Google.\n\nArgs:\n    request: FastAPI request (used to compute the callback URL).\n    return_to: Optional post-callback UI path (must be an internal\n        absolute path, e.g. ``/settings/accounts``).\n    session_cookie: Existing signed session, when the caller is\n        connecting another mailbox to the same user.\n    settings: Cached :class:`Settings`.\n\nReturns:\n    A 302 redirect to Google's consent screen.",
        "operationId": "gmail_oauth_start_api_v1_oauth_gmail_start_get",
        "parameters": [
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          },
          {
            "name": "briefed_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oauth/gmail/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Google OAuth callback \u2014 exchange code, persist account",
        "description": "Handle the Google consent-screen redirect.\n\nArgs:\n    request: FastAPI request.\n    code: Authorization code from Google.\n    state: State parameter Google echoed back.\n    oauth_state_cookie: Pre-authorize cookie set by :func:`start`.\n    settings: Cached :class:`Settings`.\n\nReturns:\n    A 302 redirect back to the UI (``return_to`` or ``/``).\n\nRaises:\n    HTTPException: 400 when the state mismatch / cookie missing.",
        "operationId": "gmail_oauth_callback_api_v1_oauth_gmail_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "briefed_oauth_state",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Oauth State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}